Change value in example configuration
The effect of using `true` for `recalculateLimitIfAlreadySet` is more
apparent than `false`, and almost a requirement for
`updateLimitsOnInterval` to work as expected. Setting the value to
default to `true` in the example configuration may help new users avoid
this common pitfall, while leaving the python default if there is no
JSON key to `True` will preserve the existing behavior for current
users.

Resolves #28